+++ cardano-slotting/testlib/Test/Cardano/Slotting/Arbitrary.hs
@@ -1,12 +1,30 @@
 {-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
+{-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE StandaloneDeriving #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Test.Cardano.Slotting.Arbitrary () where
 
-import Cardano.Slotting.Slot
+import Cardano.Slotting.Block (BlockNo (..))
+import Cardano.Slotting.Slot (
+  EpochInterval (..),
+  EpochNo (..),
+  EpochSize (..),
+  SlotNo (..),
+  WithOrigin (..),
+ )
+import Cardano.Slotting.Time (SystemStart (..))
 import Test.QuickCheck
+import Test.QuickCheck.Instances.Time ()
+
+deriving instance Arbitrary BlockNo
+
+deriving instance Arbitrary EpochNo
+
+deriving instance Arbitrary EpochSize
+
+deriving instance Arbitrary EpochInterval
 
 instance Arbitrary SlotNo where
   arbitrary =
@@ -19,4 +37,10 @@ instance Arbitrary SlotNo where
 
   shrink (SlotNo n) = [SlotNo n' | n' <- shrink n, n' > 0]
 
-deriving newtype instance Arbitrary EpochNo
+instance Arbitrary t => Arbitrary (WithOrigin t) where
+  arbitrary = frequency [(20, pure Origin), (80, At <$> arbitrary)]
+  shrink = \case
+    Origin -> []
+    At x -> Origin : map At (shrink x)
+
+deriving instance Arbitrary SystemStart